feat(ui): surface programs on landing, move project search into program pages
The landing page now shows only the hero, live stats, and program cards
(heading "Explore past programs"); the featured unit and the projects
filter/grid are gone. The search view (search box, WINNERS/ALL toggle,
category filters, grid + detail modal) is extracted into a reusable
ProjectExplorer and rendered on each program's detail page, scoped to
that program's entries.
